🔨refactor: BankAccountApplication の amount の型を Entity の内部実装に合わせて BigInt に変更する

+++ app/adapter/src/main/scala/myapp/adapter/account/BankAccountApplication.scala
@@ -10,12 +10,12 @@ trait BankAccountApplication {
   def deposit(
       accountNo: AccountNo,
       transactionId: TransactionId,
-      amount: Int,
+      amount: BigInt,
   )(implicit appRequestContext: AppRequestContext): Future[BigInt]
 
   def withdraw(
       accountNo: AccountNo,
       transactionId: TransactionId,
-      amount: Int,
+      amount: BigInt,
   )(implicit appRequestContext: AppRequestContext): Future[BigInt]
 }
